clean up & restructure
