+++ src/types/transaction.rs
@@ -1,12 +1,8 @@
-use std::time::Duration;
-
 use holochain_types::dna::ActionHashB64;
-use holochain_types::dna::AgentPubKey;
 use holochain_types::dna::AgentPubKeyB64;
 use holochain_types::dna::EntryHashB64;
 use holochain_types::prelude::CapSecret;
 use holochain_types::prelude::Timestamp;
-use holofuel_types::fuel::Fuel;
 
 #[derive(serde::Serialize, serde::Deserialize, Debug)]
 pub enum AcceptedBy {
@@ -85,38 +81,4 @@ pub struct InvoiceNote {
     // disk_usage_logs_range: Vec<ActionHashB64>,
     #[serde(flatten)]
     pub invoiced_items: InvoicedItems,
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize, Debug)]
-pub struct JurisdictionAndCategoryPreferences {
-    pub value: Vec<String>,
-    pub is_exclusion: bool,
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Debug, Clone)]
-pub struct HostingPreferences {
-    pub max_fuel_before_invoice: Fuel,
-    pub price_compute: Fuel,
-    pub price_storage: Fuel,
-    pub price_bandwidth: Fuel,
-    pub max_time_before_invoice: Duration,
-    pub invoice_due_in_days: u8,
-    pub jurisdiction_prefs: Option<JurisdictionAndCategoryPreferences>,
-    pub categories_prefs: Option<JurisdictionAndCategoryPreferences>,
-}
-
-#[derive(serde::Serialize, serde::Deserialize, Debug)]
-pub struct ServiceloggerHappPreferences {
-    pub provider_pubkey: AgentPubKey,
-    pub max_fuel_before_invoice: Fuel,
-    pub price_compute: Fuel,
-    pub price_storage: Fuel,
-    pub price_bandwidth: Fuel,
-    pub max_time_before_invoice: Duration,
-    pub invoice_due_in_days: u8, // how many days after an invoice is created it it due
-}
-
-pub struct PublisherJurisdiction {
-    pub happ_id: ActionHashB64,
-    pub jurisdiction: Option<String>,
 }
