+++ src/utils/hbs.rs
@@ -6,15 +6,43 @@ use hpos_hc_connect::hha_agent::HHAAgent;
 use hpos_hc_connect::hpos_agent::get_hpos_config;
 use reqwest::Response;
 use serde::{Deserialize, Serialize};
+use tracing::warn;
 
-#[derive(Debug, Deserialize)]
-pub struct HostingCriteria {
+#[derive(Debug, Deserialize, Clone)]
+pub struct HostCredentials {
     #[allow(dead_code)]
     pub id: Option<String>,
     pub jurisdiction: Option<String>,
-    pub kyc: KycLevel,
+    pub kyc: Option<KycLevel>,
+    // #[serde(rename = "camel_case")]
+    // pub public_key: Option<String>,
+    // pub email: String,
 }
-#[derive(Debug, Serialize, Deserialize, PartialEq)]
+impl HostCredentials {
+    pub fn is_host_in_valid_jurisdiction(
+        &self,
+        should_exclude_happ_jurisdictions: bool,
+        happ_jurisdictions: &[String],
+    ) -> bool {
+        let host_jurisdiction = match self.jurisdiction.to_owned() {
+            Some(j) => j,
+            None => {
+                warn!("Host's jurisdiction not available");
+                return false;
+            }
+        };
+        if should_exclude_happ_jurisdictions {
+            // If the host jurisdiction is present in the list that the hApp Manager has used,
+            // then the host jurisdiction is invalid
+            !happ_jurisdictions.contains(&host_jurisdiction)
+        } else {
+            // Otherwise, the host jurisdiction is valid if it exists in the happ's list of jurisdictionss
+            happ_jurisdictions.contains(&host_jurisdiction)
+        }
+    }
+}
+
+#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
 pub enum KycLevel {
     #[serde(rename = "holo_kyc_1")]
     Level1,
@@ -29,26 +57,27 @@ impl HbsClient {
         let client = reqwest::Client::builder().build()?;
         Ok(Self { client })
     }
-    pub async fn get_hosting_criteria(&self) -> Option<HostingCriteria> {
+    pub async fn get_host_hosting_criteria(&self) -> Option<HostCredentials> {
         match self.get_access_token().await {
             Ok(v) => v,
             Err(e) => {
                 tracing::warn!("Unable to get kyc & jurisdiction: {:?}", e);
                 tracing::warn!("returning default kyc level 1");
                 tracing::warn!("returning default jurisdiction of None");
-                Some(HostingCriteria {
+                Some(HostCredentials {
                     id: None,
                     jurisdiction: None,
-                    kyc: KycLevel::Level1,
+                    kyc: Some(KycLevel::Level1),
                 })
             }
         }
     }
 
-    async fn get_access_token(&self) -> Result<Option<HostingCriteria>> {
+    async fn get_access_token(&self) -> Result<Option<HostCredentials>> {
         let response = self.inner_get_access_token().await?;
         tracing::debug!("response received");
         let mut body = response.text().await?;
+
         // 504 Gateway Timeout
         // here we either need to retry once more or end the script
         if body.contains("error code: 504") {
@@ -60,10 +89,11 @@ impl HbsClient {
                 return Ok(None);
             }
         }
+
         tracing::debug!("Result: {}", body);
         let result: serde_json::Value = serde_json::from_str(&body)?;
-        let h: HostingCriteria = serde_json::from_value(result)?;
-        tracing::debug!("HostingCriteria: {:?}", h);
+        let h: HostCredentials = serde_json::from_value(result)?;
+        tracing::debug!("HostCredentials: {:?}", h);
         Ok(Some(h))
     }
 
@@ -77,7 +107,9 @@ impl HbsClient {
 
         let mut core_app = HHAAgent::spawn(None).await?;
         let pub_key = core_app.pubkey().await?;
+
         tracing::debug!("email: {:?}, pub_key: {:?}", email, pub_key);
+
         #[derive(Serialize, Deserialize, Debug, PartialEq, Clone, SerializedBytes)]
         #[allow(non_snake_case)]
         struct Body {
@@ -100,6 +132,7 @@ impl HbsClient {
                     .into(),
             )
             .await?;
+
         tracing::debug!("Signature: {:?}", signature);
 
         let connection = Self::connect()?;
